Refactor user type definitions and remove obsolete auth store files
